+++ src/Sylius/Bundle/SalesBundle/DependencyInjection/SyliusSalesExtension.php
@@ -39,13 +39,13 @@ class SyliusSalesExtension extends Extension
         $driver = $config['driver'];
 
         if (!in_array($driver, SyliusSalesBundle::getSupportedDrivers())) {
-            throw new \InvalidArgumentException(sprintf('Driver "%s" is unsupported for this extension.', $config['driver']));
+            throw new \InvalidArgumentException(sprintf('Driver "%s" is unsupported for SyliusSalesBundle.', $driver));
         }
 
         $loader->load(sprintf('driver/%s.xml', $driver));
 
         $container->setParameter('sylius_sales.driver', $driver);
-        $container->setParameter('sylius.engine', $config['engine']);
+        $container->setParameter('sylius_sales.driver.'.$driver, true);
 
         $this->mapClassParameters($config['classes'], $container);
         $this->mapValidationGroupParameters($config['validation_groups'], $container);