+++ src/Sylius/Bundle/SalesBundle/CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
 
 * Symfony 2.3 upgrade.
 * Removed ``Entity`` classes, please extend ``Model``.
+* Removed sellable interface and order builder to give more freedom for custom implementations.
 
 ### v0.1.0